add array_format_time_util.rb
